gogga-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
cerebras_cloud_sdk>=1.0.0
httpx>=0.26.0
python-multipart>=0.0.9
sqlmodel>=0.0.14
python-dotenv>=1.0.1
email-validator>=2.1.0
asyncpg>=0.29.0
alembic>=1.13.0
psutil>=5.9.0
pytz>=2024.1
# optillm runs in its own container (cepo service), not needed here
posthog>=7.0.0
apscheduler>=3.10.0
# Math tools dependencies
scipy>=1.14.0
numpy>=2.0.0
# GoggaTalk - Google AI Live API for voice chat (not Vertex AI - Live API not available in Vertex SDK)
google-genai>=1.0.0
